Moved new event framework to experimental. Added basic public API doc. (#115)
* Basic documentation of provided events.

* Moved new event framework to own module. New event framework exported under experimental, old framework exported by default.

* Renamed event test files to match project naming convention.
