+++ src/experimental/event/test/interactive/.eslintrc
@@ -1,0 +1,8 @@
+{
+  "rules": {
+    "no-console": 0
+  },
+  "globals": {
+    "console": true
+  }
+}
