+++ src/experimental/event/test/interactive/index.js
@@ -1,0 +1,3 @@
+import renderTestSelectionPage from './render-test-selection-page';
+
+renderTestSelectionPage();
