+++ src/experimental/event/test/interactive/test-pages/legacy-mouse-click.js
@@ -1,10 +1,10 @@
 /* global document */
 
-import {addEvents as addEventsLegacy} from '../../../src/core/eventLegacy';
-import CrosshairCanvas from '../CrosshairCanvas';
+import {addEvents as addEventsLegacy} from '../../../../../core/event';
+import CrosshairCanvas from '../crosshair-canvas';
 
 export default {
-  name: '(legacy) mouse click (document scroll, no cache)',
+  name: '(legacy) mouse click (document scroll, cache enabled)',
   render(testArea) {
     const canvasElement = document.createElement('canvas');
     canvasElement.height = 2 * document.documentElement.clientHeight;
@@ -16,9 +16,7 @@ export default {
         console.log(eventInfo);
         crosshairCanvas.setCrosshair({key: 'click', position: {x: eventInfo.x, y: eventInfo.y}});
       },
-      centerOrigin: false,
-      cachePosition: false,
-      cacheSize: false
+      centerOrigin: false
     });
   }
 };
