+++ src/experimental/event/test/interactive/test-pages/new-mouse-click.js
@@ -1,22 +1,21 @@
 /* global document */
 
-import {addEvents as addEventsLegacy} from '../../../src/core/eventLegacy';
-import CrosshairCanvas from '../CrosshairCanvas';
+import addEvents from '../../../add-events';
+import CrosshairCanvas from '../crosshair-canvas';
 
 export default {
-  name: '(legacy) mouse click (document scroll, cache enabled)',
+  name: '(new) mouse click (document scroll)',
   render(testArea) {
     const canvasElement = document.createElement('canvas');
     canvasElement.height = 2 * document.documentElement.clientHeight;
     canvasElement.width = 2 * document.documentElement.clientWidth;
     testArea.appendChild(canvasElement);
     const crosshairCanvas = new CrosshairCanvas(canvasElement);
-    addEventsLegacy(canvasElement, {
-      onClick: (eventInfo) => {
+    addEvents(canvasElement, {
+      onClick(eventInfo) {
         console.log(eventInfo);
-        crosshairCanvas.setCrosshair({key: 'click', position: {x: eventInfo.x, y: eventInfo.y}});
-      },
-      centerOrigin: false
+        crosshairCanvas.setCrosshair({key: 'click', position: eventInfo.pointerPosition});
+      }
     });
   }
 };
