+++ src/experimental/event/test/interactive/test-pages/new-touch-gesture-more-features.js
@@ -1,17 +1,7 @@
 /* global document */
 
-import {addEvents} from '../../../src/core/event';
-import CrosshairCanvas from '../CrosshairCanvas';
-
-function renderPinch(eventInfo, crosshairCanvas) {
-  crosshairCanvas.setCrosshair({
-    key: 'focalPoint',
-    position: eventInfo.focalPoint,
-    sizePx: 50*eventInfo.scale,
-    color: 'black',
-    style: 'circleBounded'
-  });
-}
+import addEvents from '../../../add-events';
+import CrosshairCanvas from '../crosshair-canvas';
 
 export default {
   name: '(new) touch gesture (more features)',
