--- test/interaction/index.js
@@ -1,3 +1,0 @@
-import renderTestSelectionPage from './renderTestSelectionPage';
-
-renderTestSelectionPage();
