--- test/interaction/testPages/index.js
@@ -1,25 +1,0 @@
-import legacyMouseClick from './legacyMouseClick';
-import legacyMouseClickNoCache from './legacyMouseClickNoCache';
-import newMouseClick from './newMouseClick';
-import newMouseClickAbsolute from './newMouseClickAbsolute';
-import newMouseClickFixed from './newMouseClickFixed';
-import newMouseDownUp from './newMouseDownUp';
-import newMouseOverMoveOut from './newMouseOverMoveOut';
-import newMouseEnterMoveLeave from './newMouseEnterMoveLeave';
-import newTouch from './newTouch';
-import newTouchGesture from './newTouchGesture';
-import newTouchGestureMoreFeatures from './newTouchGestureMoreFeatures';
-
-export default [
-  legacyMouseClick,
-  legacyMouseClickNoCache,
-  newMouseClick,
-  newMouseClickAbsolute,
-  newMouseClickFixed,
-  newMouseDownUp,
-  newMouseOverMoveOut,
-  newMouseEnterMoveLeave,
-  newTouch,
-  newTouchGesture,
-  newTouchGestureMoreFeatures
-];
